+++ client/app/shopping-cart/_components/cart-title/cart-title.module.scss
@@ -1,6 +1,6 @@
 .j-shoppingCartTitleBox .j-shoppingCartTitle {
     font-family: Inter;
-    font-size: 64px;
+    font-size: 54px;
     font-style: normal;
     font-weight: 500;
     line-height: normal;
