+++ client/app/shopping-cart/_components/checkout-form-step3/shopping-cart-step3.module.scss
@@ -2,14 +2,14 @@
     color: #000;
     text-align: center;
     font-family: "Noto Sans";
-    font-size: 48px;
+    font-size: 30px;
     font-style: normal;
     font-weight: 500;
     line-height: normal;
   }
   .j-addressDetailContent{
     font-family: "Noto Sans";
-    font-size: 11px;
+    font-size: 12px;
     font-style: normal;
     font-weight: 500;
     line-height: normal;
