+++ client/app/shopping-cart/cart-empty/shopping-cart-empty.scss
